requirements.txt
llama-cpp-python
transformers
torch
torchvision
torchaudio
huggingface_hub